requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
opencv-python>=4.9.0
pillow>=10.2.0
pytest>=8.0.0
tenacity>=8.2.3 